ENH: AtlasCreator: Use Double as re-cast type instead of Float. Remove all used temporary files during cleanup stage. Fixed overflow problem in connection with dynamic registration.
git-svn-id: http://svn.slicer.org/Slicer3/trunk@16086 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Modules/AtlasCreator/AtlasCreatorGUI.py
@@ -135,7 +135,7 @@ class AtlasCreatorGUI(ScriptedModuleGUI):
         
         if normalize:
             self._outputCastCombo.SetEnabled(0)
-            self._outputCastCombo.GetWidget().SetValue("Float")
+            self._outputCastCombo.GetWidget().SetValue("Double")
         else:
             self._outputCastCombo.SetEnabled(1)
